+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://semver.org/spec/v2.0.0.html), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added a Settings page (link in the top bar, next to "Go Race!" and "Career") where you can customize various aspects of your AMS2 Career experience ([#11](https://github.com/abesto/ams2-career/issues/11))
+  - Cross-Discipline XP gains can be disabled ([#48](https://github.com/abesto/ams2-career/issues/48))
+
+### Fixed
+
+- Achievements contained an optional function for formatting their current and maximum progress numbers. This was then stored in Redux state, which is not a valid operation. This didn't lead to user-facing bugs (yet), but it did generate warnings in the console. Refactored the relevant code to avoid this issue.
+
 ## [1.1.0] - 2022-06-08
 
 ### Added