+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://semver.org/spec/v2.0.0.html), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.2.0] - 2022-06-11
 
 ### Added
 
@@ -30,5 +30,6 @@ The format is based on [Keep a Changelog](https://semver.org/spec/v2.0.0.html), 
 
 - We now have a changelog! The app remembers what version was last seen, and only displays new items. Incidentally, this item will not be shown to any users, because the first time a changelog-enabled app version is loaded, we set the latest version as "seen". After all, a changelog makes no sense for a brand-new user in the general case, and the initial rollout is a one-off event.
 
+[1.2.0]: https://github.com/abesto/ams2-career/compare/v1.1.0..v1.2.0
 [1.1.0]: https://github.com/abesto/ams2-career/compare/v1.0.0..v1.1.0
 [1.0.0]: https://github.com/abesto/ams2-career/releases/tag/v1.0.0